\name{as.data.frame}
\alias{as.data.frame}

\title{ Get the data.frame from an epx object. }
\description{
S3 method. Get the data.frame of an epx object created by read.epx
}
\usage{
as.data.frame(x)
}

\arguments{
  \item{x}{
\bold{Object} - an epx object created by read.epx
}
}
\details{
----
}

\value{
----

}
\references{
----
}
\author{
Jean Pierre Decorps <jp.decorps@epiconcept.fr>
}
\note{
----
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
----
}
\examples{
# -----------
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~documentation }
